feat(vibe): project domain binds, DNS verify, Caddy TLS (#756 #774 #770)
- ProjectDomains: project_domains table (bind/list/unbind/verify_dns/issue_tls),
  TXT ownership check via dig (_gb-verify.<domain>), Caddy ACME route upsert
- caddy.rs: admin API driver (route id gbd-{domain}, dial {container}.incus:80)
- domains_api.rs: REST surface (bind/list/unbind/verify/tls)
- domains_tool.rs: LLM tools domain/bind|verify|tls (selection via state.db_pool)
- publish.rs: publish auto-binds optional domain; fixed E0425 repo->repo_name
- harness: allow dig/nslookup; fix git clone arg + byte-limit type
- UI: new-project/deploy set currentProjectId and bind external domain after deploy
- fixes: diesel into_boxed pg::Pg, Nullable<Text> for Option columns, QueryableByName
  UUID fields, projects.rs dynamic-bind list/get, vms_api Json() wrap, botautotask

+++ botserver/crates/botvibe/src/harness/mod.rs
@@ -67,13 +67,13 @@ pub fn ensure_workspace(project: &str) -> Result<PathBuf, String> {
     Ok(root)
 }
 
-pub fn read_rel_file(project: &str, rel: &str, max_bytes: usize) -> Result<Vec<u8>, String> {
+pub fn read_rel_file(project: &str, rel: &str, max_bytes: u64) -> Result<Vec<u8>, String> {
     let path = resolve_workspace_path(project, rel)?;
     let meta = std::fs::metadata(&path).map_err(|e| format!("stat {rel}: {e}"))?;
     if meta.is_dir() {
         return Err(format!("{rel} is a directory"));
     }
-    if meta.len() > max_bytes {
+    if (meta.len() as u64) > max_bytes {
         return Err(format!("{rel} exceeds {max_bytes} bytes"));
     }
     std::fs::read(&path).map_err(|e| format!("read {rel}: {e}"))
